requirements.txt
arrow==1.4.0
asgiref==3.11.0
binaryornot==0.4.4
boto3==1.42.4
botocore==1.42.4
certifi==2025.11.12
channels==4.3.2
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
Django==6.0
django-channels==0.7.0
django-storages==1.14.6
django-tailwind==4.4.2
djangorestframework==3.16.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
nodejs==0.1.1
oauthlib==3.3.1
optional-django==0.1.0
pillow==12.0.0
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.3
qrcode==8.2
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
text-unidecode==1.3
tzdata==2025.2
urllib3==2.6.0
uvicorn==0.38.0
